+++ src/plone/formwidget/recaptcha/view.py
@@ -57,6 +57,14 @@ class RecaptchaView(BrowserView):
         return None
 
     def verify(self, input=None):
+
+        # Do not validate recaptcha on form inline validation.
+        # This automatically makes the next request (form submit) already
+        # invalid. This usually happens if the captcha is not the last field
+        # on a form.
+        if self.request.URL.endswith('z3cform_validate_field'):
+            return
+
         info = IRecaptchaInfo(self.request)
         if info.verified:
             return True